+++ R/query.R
@@ -92,10 +92,10 @@ db_bind <- function(res, params, ..., allow_named_superset) {
 setMethod("dbFetch", "SQLiteResult", function(res, n = -1, ...,
                                               row.names = pkgconfig::get_config("RSQLite::row.names.query", FALSE)) {
   row.names <- compatRowNames(row.names)
-  if (length(n) != 1) stopc("n must be scalar")
-  if (n < -1) stopc("n must be nonnegative or -1")
+  if (length(n) != 1) stopc("`n` must be scalar")
+  if (n < -1) stopc("`n` must be nonnegative or -1")
   if (is.infinite(n)) n <- -1
-  if (trunc(n) != n) stopc("n must be a whole number")
+  if (trunc(n) != n) stopc("`n` must be a whole number")
   ret <- result_fetch(res@ptr, n = n)
   ret <- convert_bigint(ret, res@bigint)
   ret <- sqlColumnToRownames(ret, row.names)